SOS-004 drum machine added, single note scheduler, tests added
