+++ src/main/scala/com/soundsofscala/transport/NoteScheduler.scala
@@ -1,9 +1,9 @@
 package com.soundsofscala.transport
 
 import cats.effect.IO
+import com.soundsofscala.models
 import com.soundsofscala.models.*
 import com.soundsofscala.models.AtomicMusicalEvent.*
-import com.soundsofscala.types.{LookAhead, NextNoteTime, ScheduleWindow, Tempo}
 import org.scalajs.dom
 import org.scalajs.dom.AudioContext
 
@@ -27,7 +27,6 @@ case class NoteScheduler(
 
   def playVoice(musicalEvent: MusicalEvent)(using audioContext: AudioContext): IO[Unit] =
     val initialNextNoteValue = NextNoteTime(audioContext.currentTime)
-    audioContext.resume()
     scheduler(musicalEvent, initialNextNoteValue) >> IO.println("Sequence finished")
 
   def scheduler(musicalEvent: MusicalEvent, nextNoteTime: NextNoteTime)(
@@ -45,20 +44,40 @@ case class NoteScheduler(
           case ScheduleState.Waiting =>
             IO.sleep(lookAheadMs.value.millis) >> scheduler(musicalEvent, nextNoteTime)
 
-      case _: AtomicMusicalEvent =>
+      case event: AtomicMusicalEvent =>
         ScheduleState(nextNoteTime) match
           case ScheduleState.Ready =>
-            for _ <- scheduleNote(nextNoteTime.value)
-            yield IO.unit
+            event match
+              case note: AtomicMusicalEvent.Note =>
+                for {
+                  _ <- scheduleNote(nextNoteTime.value, note)
+                } yield IO.unit
+              case drumStroke: AtomicMusicalEvent.DrumStroke =>
+                for {
+                  _ <- scheduleDrum(nextNoteTime.value, drumStroke)
+                } yield IO.unit
+              case AtomicMusicalEvent.Rest(_) =>
+                IO.unit
+              // TODO implement Chords
+              case AtomicMusicalEvent.Harmony(_, _) =>
+                IO.unit
           case ScheduleState.Waiting =>
             IO.sleep(lookAheadMs.value.millis) >> scheduler(musicalEvent, nextNoteTime)
 
-  def scheduleNote(time: Double)(using audioContext: AudioContext): IO[Unit] =
+  def scheduleNote(time: Double, musicEvent: Note)(using audioContext: AudioContext): IO[Unit] =
     println(s"${audioContext.currentTime}: Scheduling note to play at $time")
-    playNote(time, "resources/audio/guitar/C-guitar.wav")
+    playNote(time, musicEvent)
 
-  def playNote(time: Double, filePath: String)(using audioContext: AudioContext): IO[Unit] =
+  def scheduleDrum(time: Double, musicEvent: DrumStroke)(
+      using audioContext: AudioContext): IO[Unit] =
+    println(s"${audioContext.currentTime}: Scheduling drum to play at $time")
+    NoteScheduler.playDrum(time, musicEvent)
+
+  def playNote(time: Double, note: Note)(using audioContext: AudioContext): IO[Unit] =
     println(s"Playing note at $time")
+    val octave = note.octave.value
+    val pitch: String = note.pitch.toString
+    val filePath = s"resources/audio/piano/$pitch$octave.wav"
 
     val request = new dom.XMLHttpRequest()
     request.open("GET", filePath, true)
@@ -71,7 +90,7 @@ case class NoteScheduler(
         data,
         buffer =>
           val gainNode = audioContext.createGain()
-          gainNode.gain.value = 1
+          gainNode.gain.value = 0.5
           val sourceNode = audioContext.createBufferSource()
           sourceNode.buffer = buffer
           sourceNode.connect(gainNode)
@@ -81,3 +100,37 @@ case class NoteScheduler(
         () => println(s"Things have gone sideways for now")
       )
     IO.delay(request.send())
+
+object NoteScheduler:
+  def playDrum(time: Double, drumStroke: DrumStroke)(
+      using audioContext: AudioContext): IO[Unit] = {
+    println(s"Playing drum at $time")
+    val filePath = drumStroke.drum match
+      case models.DrumVoice.Kick => "resources/audio/drums-808/Kick808.wav"
+      case models.DrumVoice.Snare => "resources/audio/drums-808/Snare808.wav"
+      case models.DrumVoice.HiHatClosed => "resources/audio/drums-808/Hats808.wav"
+      case models.DrumVoice.Clap => "resources/audio/drums-808/Clap808.wav"
+      case _ => "resources/audio/drums-808/G.wav"
+
+    val request = new dom.XMLHttpRequest()
+    request.open("GET", filePath, true)
+    request.responseType = "arraybuffer"
+
+    request.onload = (_: dom.Event) =>
+      val data = request.response.asInstanceOf[ArrayBuffer]
+
+      audioContext.decodeAudioData(
+        data,
+        buffer => {
+          val gainNode = audioContext.createGain()
+          gainNode.gain.value = drumStroke.velocity.getNormalisedVelocity
+          val sourceNode = audioContext.createBufferSource()
+          sourceNode.buffer = buffer
+          sourceNode.connect(gainNode)
+          gainNode.connect(audioContext.destination)
+          sourceNode.start(time)
+        },
+        () => println(s"Things have gone sideways for now")
+      )
+    IO.delay(request.send())
+  }
